Updated workflow

+++ src/scripts/user-delete.ts
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import { MongoDB } from "../db/db-connection";
 import User_Test from "../models/user-model";
 
@@ -5,10 +6,14 @@ import User_Test from "../models/user-model";
 const deleteZeroScoreUsers = async () => {
   try {
     console.log("Starting deletion process...");
-
-    // Get the MongoDB connection
-    const db = MongoDB.getInstance();
-    await db.connect();
+    const databaseUri = process.env.MONGODB_URI;
+    if (!databaseUri) {
+      console.error("DATABASE_URL is not set");
+      process.exit(1);
+    }
+  
+    await mongoose.connect(databaseUri);
+    console.log("MongoDB connected successfully");
 
     // // Delete the users
     await User_Test.deleteMany({ $or: [
